feat: Add Google Analytics tracking to website
This commit introduces Google Analytics tracking to the website. The
tracking ID is set as an environment variable and is used to initialize
Google Analytics on page load. Page views are logged whenever the user
navigates to a new page.

The tracking allows for better understanding of user behavior on the
website, which can inform future improvements and feature development.

Additionally, this commit includes updates to several dependencies and
minor adjustments to the website's UI.

https://medium.com/@dilhancodes/how-to-install-google-analytics-in-react-project-in-2024-205759218057